scraping/requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chromedriver-binary==109.0.5414.74.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.31
outcome==1.2.0
protobuf==3.20.1
pycparser==2.21
PySocks==1.7.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
